lav_map/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-betterforms==2.0.0
django-cleanup==7.0.0
django-crispy-forms==1.14.0
django-environ==0.10.0
gunicorn==20.1.0
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
psycopg2-binary==2.9.6
